Created a separate document for tree for readability
